+++ applicant-frontend/src/Components/ResetPassword/resetpassword.css
@@ -1,0 +1,42 @@
+.resetpassword-wrapper{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 36px;
+    padding-bottom: 24px;
+    border: 1px solid #e8e9e9;
+    width: 475px;
+    margin-right: auto;
+    margin-left: auto;
+    box-shadow: 0 2px 5px 0 rgba(55, 161, 255, 0.26), 0 4px 12px 0
+     rgba(121, 155, 254, 0.25);
+    margin-top:2.4em;
+ }
+ h1{
+     text-align: center;
+   }
+   .resetpassword-header-image {
+     width: 300px;
+     height: 50px;
+   }
+   img {
+      border: 0;
+      vertical-align: middle;
+   }
+   .btn{
+     border-radius: 10em;
+     font-weight: 400;
+   }
+   @media screen and (max-width: 550px) {
+     .resetpassword-wrapper{
+       width: auto;
+       margin-top: 0;
+       min-height: 100vh;
+     }
+   }
+   @media screen and (max-width: 550px) {
+     .resetpassword-wrapper .resetpassword-header-image {
+       width: auto;
+       margin-top: 0;
+     }
+   }
