+++ applicant-frontend/src/Pages/Login/Login.jsx
@@ -3,13 +3,15 @@
 import React, { Component } from 'react';
 import { Button, Form, Container } from 'react-bootstrap';
 import { Redirect, Link } from 'react-router-dom';
-import Handlebars  from 'handlebars';
+// import Handlebars  from 'handlebars';
 
 import './styles.css';
 import Header from '../../Components/Header/Header.jsx';
 import Footer from '../../Components/Footer/Footer.jsx';
-import { postLogin, postForgotPassword } from '../../Providers/ApiRequests.js';
+import { postLogin } from '../../Providers/ApiRequests.js';
 import Input from '../../Components/Input/Input.jsx';
+import logo from '../../Components/Header/logo.jpg';
+
 
 const TEXTS = require('../../static.json');
 
@@ -23,7 +25,7 @@ class Login extends Component {
       'isLogged': false
     };
     this.login = this.login.bind(this);
-    this.forgotPassword = this.forgotPassword.bind(this);
+    // this.forgotPassword = this.forgotPassword.bind(this);
     this.handleChangeMail = this.handleChangeMail.bind(this);
     this.handleChangePassword = this.handleChangePassword.bind(this);
   }
@@ -48,16 +50,16 @@ class Login extends Component {
     });
   }
 
-  forgotPassword() {
-    postForgotPassword({ mail: this.state.mail.trim() })
-    .then((response) => {
-      const forgotPasswordTemplate = Handlebars.compile(TEXTS.SUCCESS_MESSAGES.FORGOT_PASSWORD);
-      this.props.handleModal(forgotPasswordTemplate({ mailAddress: this.state.mail.trim() }), 'Success');
-    })
-    .catch((error) => {
-      this.props.handleModal(error.message ? error.message : error.toString(), 'Error');
-    });
-  }
+  // forgotPassword() {
+  //   postForgotPassword({ mail: this.state.mail.trim() })
+  //   .then((response) => {
+  //     const forgotPasswordTemplate = Handlebars.compile(TEXTS.SUCCESS_MESSAGES.FORGOT_PASSWORD);
+  //     this.props.handleModal(forgotPasswordTemplate({ mailAddress: this.state.mail.trim() }), 'Success');
+  //   })
+  //   .catch((error) => {
+  //     this.props.handleModal(error.message ? error.message : error.toString(), 'Error');
+  //   });
+  // }
 
   handleChangeMail(value) {
     this.setState((prevState) => {
@@ -81,15 +83,15 @@ class Login extends Component {
             :
             <div>
               <Header/>
-              <Container  style={{
-                padding: "36px",
-                paddingBottom: "24px",
-                border: "1px solid #e8e9e9",
-                width: "475px",
-                marginRight: "auto",
-                marginLeft: "auto"
-            }}>
-                <h2>{ TEXTS.LOGIN_VIEW.TITLE }</h2>
+              <Container className="wrapper">
+                
+            
+         <h1>
+           <Link className="login-header-link" to="/login">
+              <img src={logo} alt="SHA" className="login-header-image"/>
+           </Link>
+          </h1>
+            <h2>{ TEXTS.LOGIN_VIEW.TITLE }</h2>
                 <Form>
                   <Form.Group controlId="formGroupEmail" className="input-group mb-3">
                     <Form.Label>Email address</Form.Label>
@@ -103,13 +105,10 @@ class Login extends Component {
                       className='form-control'></Input>
                   </Form.Group>
                   <Form.Group>
-                    <Button size='lg' onClick={ this.login } variant="primary" block>
+                    <Button className="submitButton" size='lg' onClick={ this.login } variant="primary" block>
                       { TEXTS.LOGIN_VIEW.SUBMIT_BUTTON }
                     </Button>
-                    {/* <Button size='lg' onClick={ this.forgotPassword } variant="dark" block>
-                      { TEXTS.LOGIN_VIEW.FORGOT_PASSWORD_BUTTON }
-                    </Button> */}
-                    <p style ={{textAlign:"end"}}>
+                    <p>
                         Forgot
                         <Link to="/forgot-password" className="text-primery">
                           {" "}
