+++ applicant-frontend/src/Pages/Login/styles.css
@@ -1,3 +1,52 @@
 .big {
   width: 100%;
 }
+.wrapper{
+   display: flex !important;
+   flex-direction: column;
+   justify-content: center;
+   padding: 36px;
+   padding-bottom: 24px;
+   border: 1px solid #e8e9e9;
+   width: 475px;
+   margin-right: auto;
+   margin-left: auto;
+   box-shadow: 0 2px 5px 0 rgba(55, 161, 255, 0.26), 0 4px 12px 0
+    rgba(121, 155, 254, 0.25);
+   margin-top:2.4em;
+}
+
+h1{
+  text-align: center;
+}
+.login-header-image {
+  width: 300px;
+  height: 50px;
+}
+img {
+  border: 0;
+  vertical-align: middle;
+}
+.submitButton{
+  background: linear-gradient(40deg, #45cafc, #303f9f) !important;
+  border-radius: 10em;
+  box-shadow: 0 5px 11px 0 rgba(85, 182, 255, 0.28), 0 4px 15px 0
+    rgba(36, 133, 255, 0.15);
+    font-weight: 400;
+}
+p{
+  text-align: end;
+}
+@media screen and (max-width: 767px) {
+  .wrapper{
+    width: auto;
+    margin-top: 0;
+    min-height: 100vh;
+  }
+}
+@media screen and (max-width: 767px) {
+  .wrapper .login-header-image {
+    width: auto;
+    margin-top: 0; 
+  }
+}
